use tinted-vim
right now, it seems tinted-nvim is still experimental. there are a
couple of plugin integration that seem to not work totally smoothly,
namely gitsigns/diff - the highlight groups are also different so i'd
like to just use the ones i have set up with rainbow delimeters for now.
i like the idea of hot reloading all nvim instances with tinty, but
reloading on window focus seems to work more consistently.